backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
passlib[bcrypt]
pyjwt
pydantic[email]
python-multipart
